Deprecate GalSimFFTSizeError (#1332)
