+++ tests/test_deprecated.py
@@ -968,6 +968,24 @@ def test_save_photons():
         assert np.allclose(np.sum(image.photons.flux), flux, rtol=0.1)
         repr(obj)
 
+@timer
+def test_galsim_fft_size_error():
+    """Test basic usage of GalSimFFTSizeError
+    """
+    # This feela a little gratuitous, since almost certainly no one used GalSimFFTSizeError
+    # for anything directly.  Even catching it seems unlikely.  But it was technically part
+    # of our API, so just deprecate it and make sure it still works appropriately.
+    err = check_dep(galsim.GalSimFFTSizeError, "Test FFT is too big.", 10240)
+    print('str = ',str(err))
+    print('repr = ',repr(err))
+    assert str(err) == ("Test FFT is too big.\nThe required FFT size would be 10240 x 10240, "
+                        "which requires 2.34 GB of memory.\nIf you can handle "
+                        "the large FFT, you may update gsparams.maximum_fft_size.")
+    assert err.size == 10240
+    np.testing.assert_almost_equal(err.mem, 2.34375)
+    assert isinstance(err, galsim.GalSimError)
+
+
 
 if __name__ == "__main__":
     runtests(__file__)
